+++ crates/sui-adapter/src/lib.rs
@@ -19,6 +19,7 @@ macro_rules! assert_invariant {
 }
 
 pub mod adapter;
+pub mod error;
 pub mod execution_engine;
 pub mod execution_mode;
 pub mod programmable_transactions;
